code restructuring for a better visualization
