+++ src/components/CardList/index.ts
@@ -1,0 +1,1 @@
+export {default as CardList} from './CardList';
