+++ src/components/ErrorMessage/index.ts
@@ -1,0 +1,1 @@
+export {default as ErrorMessage} from './ErrorMessage';
