+++ src/components/List/List.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {FlatList} from 'react-native';
 import {IListProps} from '../../utils/interfaces';
-import CardList from '../CardList';
+import {CardList} from '../CardList';
 
 const List = (props: IListProps) => {
   const {data, refreshData} = props;
