+++ src/components/TableDetail/index.ts
@@ -1,0 +1,1 @@
+export {default as TableDetail} from './TableDetail';
