+++ src/screens/Detail/index.ts
@@ -1,0 +1,1 @@
+export {default as Detail} from './Detail';
